Move lifetime methods into separate file
